filter children from parent oprions
